Move identity provider client to new Java package
