+++ config-model/src/main/java/com/yahoo/vespa/model/container/IdentityProvider.java
@@ -13,7 +13,7 @@ import com.yahoo.vespa.model.container.component.SimpleComponent;
  * @author mortent
  */
 public class IdentityProvider extends SimpleComponent implements IdentityConfig.Producer {
-    public static final String CLASS = "com.yahoo.vespa.athenz.identityprovider.AthenzIdentityProviderImpl";
+    public static final String CLASS = "com.yahoo.vespa.athenz.identityprovider.client.AthenzIdentityProviderImpl";
     public static final String BUNDLE = "vespa-athenz";
 
     private final AthenzDomain domain;
