+++ vespa-athenz/src/main/java/com/yahoo/vespa/athenz/identityprovider/client/AthenzIdentityProviderImpl.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.athenz.identityprovider;
+package com.yahoo.vespa.athenz.identityprovider.client;
 
 import com.google.inject.Inject;
 import com.yahoo.component.AbstractComponent;
